requirements.txt
# Reddit Marketing Bot - Requirements
# This file contains all the necessary dependencies for the Reddit Marketing Bot

# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
PyPDF2>=3.0.1

# AI and LangChain Dependencies
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.1
langgraph>=0.0.26

# Google AI Dependencies
google-generativeai>=0.3.0

# Reddit API
praw>=7.7.0

# Additional Utilities
python-dotenv>=1.0.0
requests>=2.31.0
asyncio-throttle>=1.0.2

# Data Processing
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Interface
plotly>=5.15.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
